+++ test/WriterPluginManagerFactoryTest.php
@@ -70,4 +70,99 @@ class WriterPluginManagerFactoryTest extends TestCase
         $writers = $factory->createService($container->reveal());
         $this->assertSame($writer, $writers->get('test'));
     }
+
+    public function testConfiguresWriterServicesWhenFound()
+    {
+        $writer = $this->prophesize(WriterInterface::class)->reveal();
+        $config = [
+            'log_writers' => [
+                'aliases' => [
+                    'test' => 'test-too',
+                ],
+                'factories' => [
+                    'test-too' => function ($container) use ($writer) {
+                        return $writer;
+                    },
+                ],
+            ],
+        ];
+
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(true);
+        $container->get('config')->willReturn($config);
+
+        $factory = new WriterPluginManagerFactory();
+        $writers = $factory($container->reveal(), 'WriterManager');
+
+        $this->assertInstanceOf(WriterPluginManager::class, $writers);
+        $this->assertTrue($writers->has('test'));
+        $this->assertSame($writer, $writers->get('test'));
+        $this->assertTrue($writers->has('test-too'));
+        $this->assertSame($writer, $writers->get('test-too'));
+    }
+
+    public function testDoesNotConfigureWriterServicesWhenServiceListenerPresent()
+    {
+        $writer = $this->prophesize(WriterInterface::class)->reveal();
+        $config = [
+            'log_writers' => [
+                'aliases' => [
+                    'test' => 'test-too',
+                ],
+                'factories' => [
+                    'test-too' => function ($container) use ($writer) {
+                        return $writer;
+                    },
+                ],
+            ],
+        ];
+
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(true);
+        $container->has('config')->shouldNotBeCalled();
+        $container->get('config')->shouldNotBeCalled();
+
+        $factory = new WriterPluginManagerFactory();
+        $writers = $factory($container->reveal(), 'WriterManager');
+
+        $this->assertInstanceOf(WriterPluginManager::class, $writers);
+        $this->assertFalse($writers->has('test'));
+        $this->assertFalse($writers->has('test-too'));
+    }
+
+    public function testDoesNotConfigureWriterServicesWhenConfigServiceNotPresent()
+    {
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(false);
+        $container->get('config')->shouldNotBeCalled();
+
+        $factory = new WriterPluginManagerFactory();
+        $writers = $factory($container->reveal(), 'WriterManager');
+
+        $this->assertInstanceOf(WriterPluginManager::class, $writers);
+    }
+
+    public function testDoesNotConfigureWriterServicesWhenConfigServiceDoesNotContainWritersConfig()
+    {
+        $container = $this->prophesize(ServiceLocatorInterface::class);
+        $container->willImplement(ContainerInterface::class);
+
+        $container->has('ServiceListener')->willReturn(false);
+        $container->has('config')->willReturn(true);
+        $container->get('config')->willReturn(['foo' => 'bar']);
+
+        $factory = new WriterPluginManagerFactory();
+        $writers = $factory($container->reveal(), 'WriterManager');
+
+        $this->assertInstanceOf(WriterPluginManager::class, $writers);
+        $this->assertFalse($writers->has('foo'));
+    }
 }